test(libtelemetry,libconfig,bridge): split visualizer-edge-cases, libconfig-getters, bridge by behaviour family
- visualizer-edge-cases.test.js (415) → -basic (visualize() Edge Cases)
  and -complex (visualize() Complex Scenarios); shared beforeEach setup
  reproduced per sibling.
- libconfig-getters.test.js (433) → -github (init/rootDir/ghToken) and
  -secrets (supabase* + msApp* getters); shared rt/mockStorageFn
  reproduced per sibling.
- bridge.test.js (413) → bridge-discussions.test.js (Load/Save/origin/
  correlation/recess reads) and bridge-dispatch.test.js (Sweep eviction,
  pending-dispatch lifecycle, concurrency); shared service fixture
  reproduced per sibling.

Assertions unchanged; test counts preserved (6, 23, 14).
Spec 0640 § C (large files), design Decision 5.
